refactor: address code review feedback
- Move SessionSnapshot type to types.ts, reuse in ipc-server
- Type MonitorEvent.event as string union
- Remove unsafe double casts (as unknown as)
- Add early return in emit() when no SSE clients connected
- Add try-catch on SSE res.write() for broken connections
- Truncate tool args before JSON.stringify for efficiency
- Record tool call errors (ok: false) in monitor
- Remove redundant comment
